Create an util package to split the logic and clean code
