+++ src/test/java/com/util/DateConverterTest.java
@@ -1,4 +1,4 @@
-package com.handlers;
+package com.util;
 
 import org.junit.jupiter.api.Test;
 import java.sql.Timestamp;
@@ -8,21 +8,21 @@ public class DateConverterTest {
 
         @Test
         public void testConvertStringToTimestamp() {
-            DateConverter dateConverter = new DateConverter();
+
             String dateString = "2024-02-15";
             Timestamp expectedTimestamp = Timestamp.valueOf("2024-02-15 00:00:00");
 
-            Timestamp actualTimestamp = dateConverter.convertStringToTimestamp(dateString);
+            Timestamp actualTimestamp = DateConverter.convertStringToTimestamp(dateString);
 
             assertEquals(expectedTimestamp, actualTimestamp);
         }
 
         @Test
         public void testConvertStringToTimestampWithInvalidDate() {
-            DateConverter dateConverter = new DateConverter();
+
             String dateString = "InvalidDate";
 
-            Timestamp actualTimestamp = dateConverter.convertStringToTimestamp(dateString);
+            Timestamp actualTimestamp = DateConverter.convertStringToTimestamp(dateString);
 
             assertNull(actualTimestamp);
         }
